Update patch set 1
Patch Set 1:

Hi, and thank you for contributing!

Adding a loop flag to FileWrapper doesn't look right to me - this is supposed to be a generic file abstraction so mixing in audio file properties is mixing up the levels of abstraction.

Also, MediaFile is not considered part of the public API (see src/native-api.md) and is not used within the WebRTC library anymore, so it is not being maintained and you should not depend on it to be around for long.

If you need this functionality I suggest you branch or replace with your own implementation.

Patch-set: 1
Reviewer: Gerrit User 5077 <5077@58829da1-049c-39fb-b951-ebdcd0984767>