Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

Not sure if we need a way for the application to piggyback an arbitrary set of messages to the sender/receiver report in a single compound packet. But if we do, that's not for this cl.

Patch-set: 4
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1