Use RTCError for internal PeerConnection methods
Calls to SetLocalDescription and SetRemoteDescription in
PeerConnection delegate to many different internal helper methods
which can fail. The error ultimately needs to propagate to the
caller and cause the SetXXXDescription to fail. Right now these
methods signal errors by returning false and copying the error
message into an out parameter. This changes these methods to
return RTCError instead and avoid the use of the out parameter.

Bug: webrtc:8587
Change-Id: Ib1d31622be742718b74780110c1bbe273d66444e