Aid triaging of bugs by printing notifications about interesting parts of an event log.
Notifications are printed for gaps in seq number, capture timestamp, arrival and send times for RTP and RTCP, and high average loss.
The notifications are printed to stderr by default, but internally they are represented as subclasses to a TriageNotification base class in order to facilitate other output formats.

Initially, this is only run if the event_log_visualizer is given the flag --print_triage_notifications.

Only the first (LOG_START, LOG_END) segment is processed.

Bug:webrtc:8383
Change-Id: If43ef7f115f622fa5552dc50951a11d5f9e3cbaa