Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5: Code-Review+1
> 
> (2 comments)
> 
> Do we have good enough test coverage for this class? Anything that should be added/improved?
> 
> LGTM after that and the comment has been considered.

Agree that the class could have better test coverage. One thing that I considered doing but in the end couldn't, was to replace the platform specific code and simply use PlatformThread. I couldn't do that however because it would change the interface of this class too much (e.g. it's possible to change the name of this class, but PlatformThread is more strict). So, that's one thing. Ultimately though, I'd prefer to use TaskQueue instead of this class.

Patch-set: 5
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0