Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)
> 
> > Patch Set 4:
> > 
> > (2 comments)
> > 
> > I also removed the UMA stat, since it affects the RmsLevel object (causing it to reset).
> 
> I don't really follow the reasoning for removing the UMA stats. Can't you just cache the previously reported UMA stats value and report that for each call to GetStats?
> 
> I definitely don't think we should remove this UMA stats since they have been there for a long time, is a clear objective metric, and is something that is actually being monitored.
> 
> Removing them would break the stats for those.

Sorry if I wasn't clear about this before, but this CL was not intended to be landed, I made it for testing purposes only (that's why I added 'COMMIT=false' in the description). If we want to actually land this I agree completely that we should not delete an existing UMA stat, and we shouldn't overwrite an existing stat in getStats either.

Patch-set: 6
Reviewer: Gerrit User 5607 <5607@58829da1-049c-39fb-b951-ebdcd0984767>