Introduce webrtc::SdpType, the chosen enum for offer/pranswer/answer
This change introduces a new enum type webrtc::SdpType to replace
cricket::ContentAction and the SessionDescriptionInterface string
types. A compatibility class is added so that
SessionDescriptionInterface::type() can be treated as an SdpType
or a string, to allow backwards and forwards compatibility.

Additionally, new overloads were added for CreateSessionDescription
to take SdpType instead of a type string.

Bug: webrtc:8613
Change-Id: I52b342f12155daf8d623646b0c21b7562f69d101