Update patch set 6
Patch Set 6: Code-Review+1

Although this is a workaround for now, is it actually reasonable to require a thread to always be started and joined from the same other thread? I guess that's usually what one wants to do, but is it always?

Patch-set: 6
Reviewer: Gerrit User 5399 <5399@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1