Preliminary structured logging in ICE using the structured data infrastructure.
This change list is dependent upon CL 24700 and implements an ICE logger
that observes various objects related to the ICE protocol. Invocation of
logging methods are examples of usage of the structured data
representation in CL 24700. The structured logs are serialized to the
text log as the sink and a placeholder interface of a file sink is also
presented, which should be rewritten once the design is consolidated
with the log backend.

Bug:
Change-Id: I5501abc868aee3026905e1239cb40b334b6b5701