Update patch set 1
Patch Set 1:

Hmm. If you call this function in debug mode, you'll get a crash, but if you call it in release mode, it'll silently do nothing. Is that acceptable? Might it be safer to crash in both cases, or is it fine to silently do nothing?

Patch-set: 1
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>