Disable coregraphics built-in mouse capture in ScreenCapturerMac
When calling CGDisplayStreamCreate(properties = nullptr) this
causes kCGDisplayStreamShowCursor to default to kCFBooleanTrue.

This CL set it to false always as it was assumed. Also if true
this causes some lags when moving the mouse pointer on the capture
side and in any case webrtc::MouseCursorMonitorMac already implements
a custom way to capture the mouse. Which appears to be more efficient
in this usecase.

Bug: webrtc:8625
Change-Id: Id0fae38fa47503d87d1890213706149762fa67fb