Update patch set 5
Patch Set 5:

(2 comments)

The reason for this change is to fix a specific iOS problem. With this in mind I could probably go with Peter's approach of exposing PortAllocator to the iOS SDK and not disturbing other clients. The problem with this approach is that the PortAllocator as far as I understand is not a part of the public api (src/webrtc/api), and thus using it from the iOS SDK would be a hack as the iOS SDK (as all the other SDKs) should only be using the public C++ API. Am I wrong here? Frankly this approach also increases the scope dramatically.

On the other hand, there's this existing option in WebRTC (PORTALLOCATOR_DISABLE_ADAPTER_ENUMERATION), which solves the iOS problem, and could be potentially useful for other clients. This is the path of least resistance. But you don't want to expose it to clients. Why is it needed then? Is it a some kind of leftover that should be removed? What are the exact reasons for not exposing it?

Justin has one argument about blocking P2P functionality which I don't quite understand, could you please clarify how going through the default network interface on iOS would block P2P? I just don't get where this side-effect is coming from, could you maybe point me on the code? Maybe I should fix that PORTALLOCATOR_DISABLE_ADAPTER_ENUMERATION doesn't disable P2P?

Please tell me what approach do you prefer, because I will be happy to go either way as long as the iOS is fixed, and it doesn't take too much time to implement, I don't wanna be getting "+1" from one of you and "-1" from another due to initial disagreement :)

Patch-set: 5
Reviewer: Gerrit User 5846 <5846@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0