Update patch set 41
Patch Set 41:

PTAL.
I noticed that for non-perf bots, we were still generating a JSON file with no perf results, and uploading it to the dashboard, so I fixed that.

Also, I changed a bit the "#if defined(WEBRTC_IOS)" statements a bit to reflect that on iOS, the code after "rtc::test::RunTestsFromIOSApp();" won't be executed.

Patch-set: 41
Reviewer: Gerrit User 5028 <5028@58829da1-049c-39fb-b951-ebdcd0984767>