Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (2 comments)
> 
> Actually, now that I think some more about this, I think we should avoid the RTP names down this low.  We might have non-RTP users of it.
> 
> Meanwhile, I also forgot about "VideoStream", which is almost exactly the same as SimulcastStream.  These really ought to be merged and called something more like SimulcastLayerConfig (and SpatialLayer could be called SpatialLayerConfig or ScalableLayerConfig or ScalableSpatialLayerConfig).
> 
> For now, we should perhaps do "Option D", which would be to simply add to the existing SimulcastStream/VideoStream structs (as badly named as they are).  If you're feeling very adventures, perhaps rename them to SimulcastLayerConfig.
> 
> SimulcastStream and VideoStream both have max bitrate already, so you'd really just need to add .active or use max bitrate == 0 (although I think I prefer .active).

I'm afraid I've opened up a can of worms here. I agree that it's best to reuse VideoStream and also that we should merge VideoStream with SimulcastStream and rename it to SimulcastLayerConfig. But I think there's some issues.

1) VideoEncoderConfig contains a video_stream_factory that has a CreateEncoderStreams function. This means the VideoEncoderConfig would both contain a vector of VideoStreams and a factory that creates VideoStreams, which seems strange to me. I'm not positive why the factory design pattern is used, but it seems like it is because you get information for creating video streams in two places. First here the video_stream_factory's member variables are set (https://cs.chromium.org/chromium/src/third_party/webrtc/media/engine/webrtcvideoengine.cc?l=1881), and then here when the width/height is known you create the vector of VideoStreams (https://cs.chromium.org/chromium/src/third_party/webrtc/video/video_stream_encoder.cc?l=571) 

A possible solution here would be creating the VideoStreams initially when the VideoEncoderConfig is created, and then the only new information you're given is the width/height. Otherwise it creates this problem of storing all this redundant information in the VideoEncoderConfig for each simulcast layer that's used to create VideoStreams later.

2) Making these types of changes - changing the VideoStream name and updating the video stream factory breaks downstream clients, so it will require multiple CLs to make these types of fixes.

What I did for now is added the field to VideoStream and then added a vector of VideoStreams to VideoEncoderConfig. An alternative to make things less wonky is add a vector of bools to VideoEncoderConfig, then use this when the VideoEncoderConfig's video_stream_factor.CreateEncoderStreams function is called to properly create the VideoStreams. The problem is then we would need to also add a vector of scale_resolution_down_by/max_bitrate/bitrate_priority when we add those as well. I think the real solution is do a refactoring of it all. I could try to tackle that but it would push back the time for implementing active for simulcast streams.

Patch-set: 6
Reviewer: Gerrit User 5800 <5800@58829da1-049c-39fb-b951-ebdcd0984767>