Remove RtcEventLoggingStarted and RtcEventLoggingStopped events.
This prevents the programmer from accidentally adding LOG_START and LOG_END events to the log without actually starting the log. This also makes it easier to ensure that the LOG_START event always end up first and the LOG_END event always ends up last in the log file.

Bug: webrtc:8111
Change-Id: I4e6c9306f8559ff184b5185f8728409f8dcebfa0