Update patch set 2
Patch Set 2:

(2 comments)

Also removed the MediaSection aliases for ContentInfo. I think it's only useful to rename them if we want to sit down and rename everything that uses ContentInfo. But that would be a fair amount of work (in WebRTC and to update downstream projects) that I'm not sure is worth it right now. And having two names for the same thing seems even more confusing, so I removed it from this CL.

Patch-set: 2
Reviewer: Gerrit User 5039 <5039@58829da1-049c-39fb-b951-ebdcd0984767>