Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I think it would make sense to add a control for disabling link-local candidates.  And implementing it seems like a nice little cleanup:  In IPIsPrivate, for IPv6 we check for IPIsLinkLocal and IPIsLoopback.  But for IPv4 we combine those two into IsPrivateV4.
> We could break that up into something like IsLinkLocalIPv4, IsLinkLocalIPv6, IsLoopbackIPv6, IsLoopbackIPv4.  
> 
> But how does the NetworkManager get the control?  I suppose we could add a port allocator flag PORTALLOCATOR_DISABLE_LINK_LOCAL and then have the PortAllocator call some kind of NetworkManager::disable_link_local_ips().

Hello,

The new approach is implemented here in a new CL:
https://webrtc-review.googlesource.com/c/src/+/36380

I haven't modified the NetworkManager, because it feels that it is responsible to enumerate all interfaces which could be working in theory, while BasicPortAllocatorSession::GetNetworks does more precise filtering to the client needs and seems more appropriate. It seems that at some point IPIsLinkLocal might be removed in IsIgnoredIPv6, and it is only there as a hack, i.e. it is a limitation of NetworkManager, not a "feature".

The rest is as discussed.
Feel free to review.

Patch-set: 5
Reviewer: Gerrit User 5846 <5846@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0