Fix reporting of perf results on video full stack tests.
For DualStreamsTest, the name of the metric reported
("dualstreams_moderately_restricted_screenshare") was repeated 4 times:
- Conference_Restricted/0
- Conference_Restricted/1
- ModeratelyRestricted_SlidesVp8_3TL_Simulcast_Video_Simulcast_High/0
- ModeratelyRestricted_SlidesVp8_3TL_Simulcast_Video_Simulcast_High/1

So only one of those tests (whichever ran last) has its metrics reported
to the perf dashboard, while the others have their metrics ignored.

I added the "/0" or "/1" as part of the metric name, to differentiate
between them.

Bug: webrtc:8566
Change-Id: I088807b66f9b7957571dccdb8fe3df0d87486bb0