Android: Move JNI helper functions to rtc_base
This CL moves the common JNI helper files to rtc_base so that it can be
re-used for the audio code. The goal is to use the JNI generation script
for the audio code. When the Android audio code is moved out from core
WebRTC, we should move out the JNI helper files as well.

Bug: webrtc:8689
Change-Id: I0beffabebc8c79c7d04b5c92e3b938292d41d6f5