Update patch set 2
Patch Set 2:

I may be biased, but I generally like to avoid this sort of use of inheritance, since the boundary between the base class and derived class tends to be fuzzy (and get fuzzier over time). A great example in our codebase is Port and its subclasses.

I *would* suggest doing this by using composition instead of inheritance (making some helper class to handle the BaseChannel interaction, I guess?), but I'm not even sure if that's worth it, given that the extra boilerplate code would exceed the current duplicate code.

But if you feel strongly, or if you have some future plans that rely on this change, feel free to go ahead with it.

Also: I think it would be preferable to make a separate CL for moving the "worker thread hop" responsibility to RtpReceiver and exposing MediaChannel to it.

Patch-set: 2
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>