Enable orphan checks for all .h files.
Now that there aren't any orphans[1], this change makes it a bit
harder to get more of them by, for instance, unlisting them from
.gn files. The previous check only covered new .h files.

The check will not catch all changes that create orphans,
such as when a file is simply dropped from a gn file. It's hard
to implement this, I believe. It should cover the major cases,
such as when a header moves between dirs.

[1] Depends on https://webrtc-review.googlesource.com/c/src/+/38200.

Bug:None
Change-Id: I6b61ea119a9ca1df6ebf381c0f5f4d8897c18b96