Reconfigure, not reconstruct, AudioReceiveStreams.
In preparation of moving ownership of voe::Channel to the audio stream classes, semantics for
changing configuration properties on the receive streams need to change, otherwise RTP, audio
decoding and NetEq state will be discarded when streams are recreated. The same pattern as for
AudioSendStream is applied, and the reconfigurable information is kept to a minimum.

AudioReceiveStream:s may still be recreated when an unsignaled stream is 'promoted' to signaled
state, and the sync label changes at the same time.

Bug: webrtc:4690
Change-Id: Ibad282965310c3c8174a91e05a659fa3e1827607