Update patch set 4
Patch Set 4:

Not lgtm. This overwrites the behavior on operating systems that are not affected by the problem and it sounds like the problem itself is not actually understood. I'd rather figure out exactly what the problem is and when addressing it, make sure that side effects aren't introduced.

As an example, an event on windows is something that is supported on all versions of windows. A condition variable is not. What binaries (dlls) provide the new functionality might be different (dependency changes), performance is likely to be different as well.

Another question is, are we already depending on stl's mutex and/or condition variable classes? If not, is there anything to consider there?

Patch-set: 4
CC: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>