Generate iOS framework umbrella header.
Instead of keeping the umbrella header in sync manually and needing
ifdefs to make it include the correct headers depending on platform,
generate it based on the headers we include in the framework target.

Can also be used to only include internal software codec headers when
compiling with support for them.

Bug: webrtc:7925
Change-Id: I63f97af1efc8710cfd62d527fcb343fed05daae2