Update patch set 1
Patch Set 1:

(8 comments)

You're listing a bunch of new files that I assume live in the Mozilla repo, mostly as added sources in existing targets, but at least once as a whole new build target. If it was a small number of files I guess it would be OK, but there are enough of them that I'm wondering if it wouldn't be better to put the info in a .gn or .gni file in the Mozilla repo and referencing that.

Patch-set: 1
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>