Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Why do you need to add this dependency to a bunch of targets that don't include the header? From your CL it looks like you really need to add it to congestion_controller and pacing only.

This is because this is only default webrtc-internal implementation. There's another in chrome in webrtc_overrides, which forwards chrome feature flags.

If I add this default implementation to congestion_controller, it will be there even in chrome, which is not what we want. Similar to field_trials_default, this new target has to be added to all standalone executables but not the modules.

Patch-set: 7
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>