Add new functionalities to JsepTransport.
Add new SetLocal/RemoteTransportDescription methods to JsepTransport.
RtcpMuxFilter and SrtpFilter is moved to the JsepTransport which is now
resposible for setting up the RtcpMux, configuring SDES parameters and
updating the encrypted extension IDs for DTLS-SRTP.

The method IsSrtpActive() is added to RtpTransportInternal and its
implementations which will be used by the BaseChannel to tell if the
SRTP is active and make sure we don't send unencrypted packets when
SRTP is required.

Bug:webrtc:8587
Change-Id: I109802ec171c13ced86a8d4f34bcb2e4dfe9d1ca