Update patch set 2
Patch Set 2:

Adding owners of core parts in WebRTC since I think they know the details and background of the initial work done for Android better.

It is not clear to me why these changes are needed given the reference to an Android specific issue. The native audio layers for desktop platforms are mainly used as test code and is not used by any major client, hence changes in those parts adds very little value as is. It is also rather complex to verify the performance given the wide range of devices out there.

Regarding changes in modules/audio_device/ I would prefer not to make any changes unless it can be shown that the changes solves an existing problem.

Patch-set: 2