Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

If there is no one using it on Mac or Linux, we should delete it. execinfo.h is definitely not a header on iOS or Android NDK.


We are using our own build system to compile WebRTC from source, we cannot compile for android when the header is not on the NDK. 

BTW, I found this fake implementation (https://android.googlesource.com/platform/external/chromium/+/master/android/execinfo.cc) on the android repo (for chromium?). Could this be the reason the android build can go through?

Patch-set: 2
Reviewer: Gerrit User 6278 <6278@58829da1-049c-39fb-b951-ebdcd0984767>