Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> > 
> > PTAL. Is it normal to have both guarded_by and pt_guarded_by with the same "guard" object?
> 
> I think we usually just use guarded_by in that case, since there's no way to get to the pointee without touching the pointer first.

For locks, guarded_by says that it's valid to get the lock, read the pointer into a local variable, release the lock, and then access the pointee. pt_guarded_by makes it clear that such use is invalid (at least to the human, not sure if the compiler tracks this property as the pointer is copied around, and complains about it).

For task queues as guards, this is maybe less likely problem, but, e.g.

  Foo* foo RTC_GUARDED_BY(&tq);

  ... running on tq ...
  Foo* copy_foo = foo;
  other_tq->PostTask([copy_foo]() {... });

should be valid use, but invalid it foo is also marked RTC_PT_GUARDED_BY.

Patch-set: 2
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>