Make corevideo_frame_buffer wrap RTCCVPixelBuffer
The corevideo_frame_buffer class duplicates a lot of the functionality
in RTCCVPixelBuffer but misses some features, like support for other
pixel formats. This CL makes corevideo_frame_buffer wrap the newer
Obj-C class to remove this duplication.

It also removes dependencies on the corevideoframebuffer_objc target
since it is no longer used internally in WebRTC, but it is still
used by downstream projects.

Bug: webrtc:8470
Change-Id: Ieae265748bab5d21fe0ee2f19993f9300a523973