Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > I have the feeling that //api:libjingle_peerconnection_api is too big and that it is too easy to depend on it, causing these circular dependencies.
> 
> For applications, it makes sense to depend on one big target with the complete webrtc or peerconnection api. But we can't depend on a target like that internally, whenever we need to use any single api header file.
> 
> To me, it would reasonable to have small *_api targets, which is what we use internally. And to provide one large target for applications, defined as just a public_deps on a collection of smaller targets. But we're also trying to avoid public_deps, so is there some other mechanism to group a set of targets and treat them as one?

I don't think there is such mechanism but in this case I am not 100% against public_deps.
If the target using public_deps is never used internally but only by external clients I am fine with it. We should document this carefully but the idea of having something like that is not a bad idea. Let me think about it, I want to be sure there isn't another way to achieve this.

Patch-set: 3
Reviewer: Gerrit User 5087 <5087@58829da1-049c-39fb-b951-ebdcd0984767>