Allow the MediaChannel to pass the parsed RTP packet to Call object.
In this CL, the BaseChannel is reponsible for parsing the received RTP
packets so that it can use the RtpDemuxer in the future. (This logic would
be moved further down to RtpTransport when the RtpDemuxer is shared
between BaseChannels).

To avoid parsing the packets for multiple times, the
MediaChannel::OnPacketReceived method is modified to take the
RtpPacketReceived as an argument and directly pass the parsed packet to
the Call object.

predecessor CL:https://webrtc-review.googlesource.com/c/src/+/46863

Bug:webrtc:8587
Change-Id: I3309bd0335238cac0d659c35e7de76b1a650ba67