Update patch set 1
Patch Set 1:

Submitting a patch for 8258.   The same RuntimeException is catch in other places in Camera1Session and Camera1Enumerator.

I was able to reproduce it locally (forcing Camera1 usage and calling start/stop Camera multiple times) and after this patch the Exception was still happening but it was catch and the Camera was recreated properly.

Patch-set: 1
Reviewer: Gerrit User 5098 <5098@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 6000 <6000@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false