Adding simulcast/SVC RD perf tests.
The purpose of having these tests is to make simulcast/SVC evaluation
more consistent and reproducible by using fixed codec/network settings
and input video. These tests are not supposed to be run on bots.

I modified VideoProcessorIntegrationTest such that it provides stats
of spatial and temporal layers to caller if it passes stats container
to ProcessFramesAndMaybeVerify.

Bug: webrtc:8524
Change-Id: I0be857e36ddcbd381d8574563272a67df4f89ccb