Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> Rebased after deletion of RTC_ACCESS_ON.
> 
> Karl, do you understand the compilation failure if we use RTC_DCHECK_RUN_ON inside a method which is annotated as RTC_RUN_ON?

I think so. RTC_ACCESS_ON is a function attribute that requires a mutex (or similar) to already be held when you call that function. RTC_DCHECK_RUN_ON expands to code that tries to acquire the mutex for the duration of the current scope.

In other words, you don't need to dynamically check that you're on encoder_queue_ by calling RTC_DCHECK_RUN_ON in that function, because the RTC_ACCESS_ON annotation statically ensures that you are on encoder_queue_.

Patch-set: 5
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>