Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> Looks good to me; however, I have two questions.
> 
> 1. I noticed that the comment "// 96 to 127 dynamic payload types for video codecs" was not included in the new code.  Was such your intent?

Yes, that was the intent. I didn't think it added anything, and we are trying to decouple the codec configuration from the RTP configuration anyway.

> 
> 2. In the new code there's a line, settings->targetBitrate = 0;. I do not see a corresponding line in the old code. Is this new line necessary?

Not strictly necessary, but I thought it was nicer to explicitly set the field rather than relying on default initialization.

Patch-set: 4
Reviewer: Gerrit User 5023 <5023@58829da1-049c-39fb-b951-ebdcd0984767>