Update patch set 7
Patch Set 7:

Fredrik, can you shed some light on how AudioMixer relates to CreatePeerConnectionFactory? chromium calls

  pc_factory_ = webrtc::CreatePeerConnectionFactory(
      worker_thread_ /* network thread */, worker_thread_, signaling_thread_,
      audio_device_.get(), CreateWebrtcAudioEncoderFactory(),
      CreateWebrtcAudioDecoderFactory(),
      cricket::ConvertVideoEncoderFactory(std::move(encoder_factory)),
      cricket::ConvertVideoDecoderFactory(std::move(decoder_factory)),
      nullptr /* audio_mixer */, nullptr /* audio_processing */);

The null audio mixer is converted to a rtc::scoped_refptr for the call. And that requires a complete declaration of AudioMixer, otherwise the compilation fails with an attempt to use AddRef and Release on an incomplete type. That is a bit silly. But I don't want to include api/audio/audio_mixer.h in the chromium code if it isn't actually using any mixer. It's more appropriate to add the include to peerconnectioninterface.h, but I can't do that easily, since that results in a circular dependency.

It's not clear to me what chromium gets, will it get no audio mixer, or a default audio mixer?

It would make some sense to delete the null arguments in the call. There is an almost matching overload  with two arguments less, but (1) that one is deprecated since it uses raw pointers instead of unique_ptr for the video encoder and decoder factories, and (2) it's unclear whether or not it will create a default audio mixer.

I could add yet another overload to CreatePeerConnectionFactory with the right arguments. But then I'd at least need to know how to get it to create or avoid creating an audio mixer automatically, whatever is the correct behavior.

Patch-set: 7
Reviewer: Gerrit User 5077 <5077@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0